feat: implement responsive header component and adjust main padding for mobile view
