+++ app/globals.css
@@ -687,3 +687,9 @@ h2 {
 main {
   padding-top: 85px;
 }
+
+@media (max-width: 767px) {
+  main {
+    padding-top: 0px;
+  }
+}
